% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_region_filenames}
\alias{get_demo_region_filenames}
\title{Get regions filenames for demo}
\usage{
get_demo_region_filenames()
}
\value{
A vector of regions filenames
}
\description{
Get regions filenames for demo
}
\examples{
regions <- get_demo_regions()
}
