% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_rna_regions}
\alias{get_demo_rna_regions}
\title{Get demo regions}
\usage{
get_demo_rna_regions()
}
\value{
A GRangesList with two genes
}
\description{
Get demo regions
}
\examples{
regions <- get_demo_rna_regions()
}
